Created Endpoint class for testing; refactored integration test; added comments
