+++ src/main/java/edu/chris/boattrack/controller/rest/BoatController.java
@@ -42,7 +42,7 @@ class BoatController {
   // end::get-aggregate-root[]
   
 //  @GetMapping(path = "/boattestmsg", produces = MediaType.APPLICATION_JSON_VALUE)
-  @GetMapping(path="/boattrack/v0.0/boat-test-message")
+  @GetMapping(path="/boat-test-message")
 //  /boattrack/v0.0/boat-test-message
   Map testMessage() {
 	  Instant now = Instant.now();
