+++ src/main/resources/application.properties
@@ -7,4 +7,4 @@ spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 #Tried the two below to configure base URI and they did not work.  spring.mvc.servlet.path is the one that works
 #spring.data.rest.base-path=/cc
 #server.servelet.context-path=/cc
-spring.mvc.servlet.path=/boattrack/v0.0
+spring.mvc.servlet.path=/bt/v1
