+++ src/test/resources/application.properties
@@ -1,3 +1,4 @@
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
 local.server.port=8080
+spring.mvc.servlet.path=/bt/v1
